Moved Program.cs initialization to Application.cs
